+++ Gruntfile.js
@@ -5,6 +5,26 @@ exports = module.exports = (grunt) => {
   loadGruntTasks(grunt);
 
   grunt.initConfig({
+    browserify: {
+      dist: {
+        files: {
+          'web/js/engine.js': 'engine/math.js',
+        },
+        options: {
+          exclude: [
+            'engine/clients.js',
+            'engine/index.js',
+            'engine/networks.js',
+            'engine/pages.js',
+            'common-errors',
+          ],
+          transform: [
+            ['babelify', { presets: ['es2015'] }],
+            ['uglifyify', { global: true }],
+          ],
+        },
+      },
+    },
     eslint: {
       src: [
         'app.js',
@@ -13,6 +33,7 @@ exports = module.exports = (grunt) => {
         'Gruntfile.js',
         'routes/**/*.js',
         'web/js/*.js',
+        '!web/js/engine.js',
         'test/**/*.js',
       ],
     },
@@ -34,7 +55,7 @@ exports = module.exports = (grunt) => {
     },
   });
 
-  grunt.registerTask('build', ['sass']);
+  grunt.registerTask('build', ['browserify', 'sass']);
   grunt.registerTask('lint', ['eslint']);
   grunt.registerTask('test', ['lint', 'mochaTest']);
 };