+++ CHANGELOG.md
@@ -8,6 +8,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 
+## [0.8.0][0.8.0] - 2016-11-19
+
+Added:
+
+* Support client-side calculation
+* Validate data files, throwing `SpecViolationError` when necessary
+
+Changed:
+
+* Throw a `UnsupportedSpecError` if a data file is found **not** using the
+  expected spec version
+
+
 ## [0.7.0][0.7.0] - 2016-11-16
 
 Changed:
@@ -100,4 +113,5 @@ This is the very first version.
 [0.5.0]:https://github.com/forfuturellc/mmtc-ke/releases/tag/v0.5.0
 [0.6.0]:https://github.com/forfuturellc/mmtc-ke/releases/tag/v0.6.0
 [0.7.0]:https://github.com/forfuturellc/mmtc-ke/releases/tag/v0.7.0
-[Unreleased]: https://github.com/forfuturellc/mmtc-ke/compare/v0.7.0...HEAD
+[0.8.0]:https://github.com/forfuturellc/mmtc-ke/releases/tag/v0.8.0
+[Unreleased]: https://github.com/forfuturellc/mmtc-ke/compare/v0.8.0...develop